Refactoring comments code structure
