+++ app/assets/stylesheets/api/posts.scss
@@ -133,3 +133,12 @@
     font-size: 14px;
 }
 
+#post-time {
+    font-size: 10px;
+    color: rgb(142, 142, 142);
+    font-weight: 400;
+    padding-left: 15px;
+    margin-top: 10px;
+    margin-bottom: 7px;
+    letter-spacing: 0.2px;
+}
