+++ app/views/api/posts/_post.json.jbuilder
@@ -1,2 +1,3 @@
 json.extract! post, :id, :user_id, :caption, :created_at
 json.photoUrl url_for(post.photo)
+
