+++ frontend/components/posts/post_index_item.jsx
@@ -1,10 +1,60 @@
 import React from 'react';
 import PostIndexLikesContainer from '../likes/post_index_likes_container';
 import CommentsIndexContainer from '../comments/comments_index_container';
+import CommentsFormContainer from '../comments/comments_form_container';
 
 class PostIndexItem extends React.Component {
     constructor(props) {
         super(props);
+        
+        this.calculateTime = this.calculateTime.bind(this);
+    }
+
+    calculateTime(date) {
+        let now = new Date();
+        let seconds = Math.floor((now - new Date(date)) / 1000);
+
+        let years = Math.floor(seconds / 31536000);
+
+        if (years === 1) {
+            return years + ' YEAR';
+        } else if (years > 1) {
+            return years + ' YEARS'
+        };
+
+        let months = Math.floor(seconds / 2592000);
+
+        if (months === 1) {
+            return months + ' MONTH';
+        } else if (months > 1) {
+            return months + ' MONTHS'
+        };
+
+        let days = Math.floor(seconds / 86400);
+
+        if (days === 1) {
+            return days + ' DAY';
+        } else if (days > 1) {
+            return days + ' DAYS'
+        };
+
+        let hours = Math.floor(seconds / 3600);
+
+        if (hours === 1) {
+            return hours + ' HOUR';
+        } else if (hours > 1) {
+            return hours + ' HOURS'
+        };
+
+        let minutes = Math.floor(seconds / 60);
+
+        if (minutes === 1) {
+            return minutes + ' MINUTE';
+        } else if (minutes > 1) {
+            return minutes + ' MINUTES'
+        };
+
+        return seconds + ' SECONDS';
     }
 
 
@@ -30,6 +80,8 @@ class PostIndexItem extends React.Component {
                             <p id="post-caption-body">{this.props.post.caption}</p>
                         </div>
                         <CommentsIndexContainer post={this.props.post} postId={postId} />
+                        <p id="post-time">{this.calculateTime(this.props.post.createdAt)} AGO</p>
+                        <CommentsFormContainer post={this.props.post} postId={postId}/>
                     </div>
                 </li>
             </div>
