+++ frontend/components/posts/post_show.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import PostIndexLikesContainer from '../likes/post_index_likes_container';
+import CommentsIndexContainer from '../comments/comments_index_container';
 
 class PostShow extends React.Component {
     constructor(props) {
@@ -8,7 +9,9 @@ class PostShow extends React.Component {
 
     componentDidMount() {
         window.scrollTo(0,0);
-        this.props.fetchPost(this.props.match.params.postId)
+        this.props.fetchPost(this.props.match.params.postId);
+        this.props.fetchAllLikes();
+        this.props.fetchAllComments();
     }
 
     componentDidUpdate(prevProps) {
@@ -35,6 +38,9 @@ class PostShow extends React.Component {
                         </div>
                         <div id="post-likes-box">
                             <PostIndexLikesContainer post={this.props.post} postId={this.props.post.id} />
+                        </div>
+                        <div>
+                            <CommentsIndexContainer post={this.props.post} postId={this.props.post.id} />
                         </div>
                     </div>
                 </div>
